pyproject.toml
[project]
name = "myer-pim-sdk"
version = "0.1.0"
description = "Python library for simplifying the interaction with Myer's PIM API (Akeneo's Product management API)"
authors = [
    {name = "DuneRaccoon",email = "benjamincsherro@hotmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "leaky-bucket-py (>=0.1.3,<0.2.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
setuptools = "^80.9.0"

